+++ packages/ui-kit/stories/components/inputs/date-picker/date-picker.stories.tsx
@@ -1,6 +1,6 @@
 import { Meta } from '@storybook/react'
 
-import { DatePicker } from '../../../../src'
+import { DatePicker } from '../../../../src/components/date-picker'
 
 export default {
   title: 'Components/Inputs/Date Picker',